(Bugfix)|Replace test bundle resources with base64-encoded resources
Swift Package Manager has no concept of a bundle or an Xcode Project, so this replaces resources linked to the test targets with base64-encoded files that are lazy-loaded into Data structs. Resources that were previously accessed via `Bundle(for:...)` are now accessed via `TestBundle.myResource`.

Additionally, I found that one of the CertificateBundle initializers that should've been public was still marked as internal.