(Feature)|Configure project for Travis CI (#8)
After some heavy back-and-forth with Travis builds, it appears that all of the intermittent issues with builds and tests have been resolved. Rather than using Travis to run the test script in a matrix, the build matrix is handled within the Rakefile.

The Rakefile is set up to run through both the current and last major OS version of each platform. For all targets except for watchOS, each test run uses the build-for-testing and the test-without-building commands to mitigate simulator timeouts (DTXProxyChannel errors) as mentioned here: travis-ci/travis-ci#6422 (comment)

Lastly, this updates OAuth2TokenDiskStoreTests.testFileStorageOperations to use a temporary directory since file permissions seem to be shaky on certain platforms/OS's on Travis, and it also updates the disk store to use safer write operations.