(Feature)|Redefine RequestSerializer protocol and update FormEncodedRequestSerializer
Historically, RequestSerializer was meant to fully serialize a URLRequest for transport. However, this is too much responsibility for something that is only meant to apply specific MIME encoding to a request, especially when query parameters will ALWAYS be form-encoded. HTTPRequestBuilder allows for specific formatting within queries, so it's been removed completely from RequestSerializer.

Additionally, FormEncodedRequestSerializer now exposes the same formatting options, but specifically for body parameters.