Define XMLTextNodeInitializable protocol (#37)
### Summary
Currently `Conduit` relies on `LosslessStringConvertible` to parse text node values out of `XMLNode`. Because this protocol is public and Swift does not allow for private extensions to public protocols, this has the downside that framework consumers trying to implement parsers for types that do not conform to this protocol, like `Int` or `Date`, will need to make those extensions also public.

To avoid this problem, this pull request implements our own custom `XMLTextNodeInitializable` protocol. While extensions for this protocol will still have to be public, these extensions will not interfere with `LosslessStringConvertible`, and only with our own custom protocol.

### Implementation
- `XMLTextNodeInitializable` protocol has been added.
- `XMLNode` value getters have been updated to conform to `XMLTextNodeInitializable`.

### Test Plan
- API interface has not changed. All unit tests are passing without any updates.