(Final Review)|Migrate to Swift 4 (#47)
This set of changes includes the full migration to Swift 4, along with structural improvements made possible through the migration. Most importantly, `BearerOAuth2Token` and `BasicOAuth2Token` have been deprecated in favor of `BearerToken` and `BasicToken`, which are struct counterparts that leverage the new `Encodable` and `Decodable` protocols, which are now preferred over the ObjC Runtime-dependent `NSCoding` protocol. In order to allow for seamless migration, storage interfaces look for a `DataConvertible` token, which is a custom protocol that's automatically implemented for structures that implement `Encodable`, `Decodable`, and `NSCoding`. This means that `BearerOAuth2Token`, `BasicOAuth2Token`, and `BearerToken` are all supported token types when handling token storage.

`BasicToken` does not conform to `DataConvertible` as there is no need for basic token storage directly within Conduit, and doing so would only enforce bad security design.