(Bugfix)|Scope should not be required for refresh token grants
In three-legged auth flows, scope is typically defined through the authorization server / authorization page. Because of this, it doesn't make sense to provide resource scope to the refresh_token grant, and stricter authorization servers will actually reject grants that contain a scope.

For the future of OAuth2RequestPipelineMiddleware, we should aim to separate this responsibility into a publicly-accessible grant strategy that can be consumed by the middleware. For now, this updates the existing logic to only set the "scope" field if one exists on the client configuration.