(Feature)|Expose token refresh strategy for custom refresh implementation and configuration (#89)
While the OAuth 2.0 spec is strict and explicit with regard to client requests and authorization server responses, the practical matter is that real-world implementations rarely follow the core spec to the tee. It's for this reason that all standard grant strategies are customizable in ways that often change from one authorization server to the next, or from one authentication extension protocol to the next. Token refreshes are a major touchpoint and place of flexibility, which is not currently accessible to the caller.

This separates the standard refresh_token grant logic into an `OAuth2RefreshTokenGrantStrategy` for consistency. Within the auth middleware, users can now provide custom refresh grants via an `OAuth2RefreshStrategyFactory` implementor, which defaults to an `OAuth2TokenRefreshGrantStrategyFactory` (provides the default refresh_token strategy). If this property is nil, then no refreshes will be performed.