(Feature)|Add Codecov for visualizing and enforcing code coverage (#91)
This sets up the initial configuration for [Codecov](https://codecov.io) support. Complementary to CI, we now have two additional tools to monitor and improve our codebase:
- Indepth coverage data and statistics on a codecov.io dashboard
- Coverage information and enforcement on every pull request

There are numerous advantages to highlighting code coverage. To name just a few:
- New code will now require passing coverage benchmarks to pass status checks
- Existing code can be audited and bolstered with better test coverage
- All source changes, no matter how big or small or "quick", will need to carry the diligence of validating tests

The current project coverage is 79%. Just to start off, I've set the target coverage to 75%, which will be raised as we continue to add more tests. The `patch` target will likely soon differ from the `project` target as well to further drive forward the importance of full coverage.