(Feature)|Add support for multi-session coordination and app containers (#93)
While Conduit technically can support session handoff between processes via file storage + group containers or via keychain groups, there is still a lack of support for concurrent processes, which makes either approach impractical. The single point of failure for this being possible is within the token refresh workflow. If two processes happen to concurrently process middleware for a unified session with an expired token, then the loser will send up a refresh token that has already been used and force both sessions to end. This experience is less than ideal.

[Interprocess communication is hard](http://ddeville.me/2015/02/interprocess-communication-on-ios-with-berkeley-sockets) outside of macOS. I've explored many different potential architectures using the few options available (primarily Darwin notifications and NSFilePresenter). After months of scrapping countless paths, I've landed on an approach that further drives Conduit's primary goal, which is simplicity. This only focuses on the target point of failure described above, and it does so using existing interfaces. To break it down:
- When a session begins a refresh, a new method on `OAuth2TokenStore` allows the middleware to lock the refresh token for a period of time
- If an alternate session hits the same code branch before the refresh completes, then it waits to observe an unlock, or until the lock expires (in case of process termination or other errors)
- When the refresh completes, the refresh token is unlocked via token storage, and a system notification is fired via `CFNotificationCenterGetDarwinNotifyCenter`. These simple notifications are dispatched through Core OS and can be observed by all processes running on a device
- The alternate session then re-processes the parameters via async recursion

As for the actual token storage, consumers are limited to unencrypted file storage, in which I've discovered a bug/flaw in design that prevents multiple OAuth2 clients, or multiple authorization levels within each client, from writing to a single location. Plain and simple, `OAuth2TokenDiskStore` was ill-defined and messy, and the expansion of state-based logic was a clear sign of too many responsibilities. This deprecates `OAuth2TokenDiskStore` in favor of `OAuth2TokenUserDefaultsStore` and `OAuth2TokenFileStore`. The new file storage requires a _directory_ instead of a _file path_, which solves the design problem. It also allows each type of storage to be focused and to scale with improvements as we see fit.