(Bugfix)|SwiftLint crashes due to Swift bug (#97)
As discussed in https://github.com/realm/SwiftLint/issues/2276, there is a SourceKit bug involving switch case annotations when string interpolation is involved, which throws off an underlying unicode conversion when running the new no_fallthrough_only rule and causes a crash. As a workaround, this moves the string interpolation outside of the switch body.