(Feature)|Add support for response middleware (#102)
Currently, Conduit only supports middleware that intercepts the request payload within a session exchange. As a reminder, the network pipeline architecture rides on the coattails of the Express.js middleware approach. One major difference between client and server networking architecture is that the server directly handles and controls the exchange from start to finish, but the client deals with the exchange in two parts. This finally adds the second part.

After a couple months of debate and iterations, I've landed on a go-forward approach for supporting response middleware. To keep changes focused and lightweight, this doesn't introduce any new patterns or models beyond the existing interfaces. This does technically change the system diagram, which will need to be updated soon afterwards.