(Bugfix) | Escape Predefined Entities in XML Serialization (#119)
### Summary
There are specific characters that are not allowed in XML without being escaped. Specifically `&`, `'`, `/`, `<`, and `>`. These are defined in the following pages of documentation -

https://www.w3.org/TR/xml/#sec-predefined-ent
https://en.wikipedia.org/wiki/List_of_XML_and_HTML_character_entity_references#Predefined_entities_in_XML

### Implementation
The escape was implemented during the serialization of each of the XMLNodes. When the value of the XMLNode is serialized, each of the Predefined Entities in XML are escaped and replaced with their escape values. 

### Test Plan
- The test `XMLRequestSerializerTests.testMalformedXMLSerializesXML()` was added to validate the escaping of each of the predefined entities. The malformed XML has each of the set characters, and validates that the XML is correct even after being escaped.