(Feature) | Include expiring token in pre-fetch hook (#125)
### Summary
Some applications might want to do additional operations before a token is refreshed. Currently, the pre-fetch hook does not include the about-to-expire token.

By passing the token refreshed in the hook, we allow applications to do additional operations that might be related to that token.

### Implementation
Pre-fetch hook now includes the token being refreshed. This token will be `nil` when the hook is called the first time a token is issued.