(Feature)|Extend cipher functionality to support hybrid encryption & SEP (#143)
This introduces two new protocols, `Encryptor` and `Decryptor`, along with a typealias `Cipher` that implements both. These are both implemented by an included `HybridCipher`, which employs hybrid encryption to encrypt/decrypt data. The hybrid keys (which are ultimately asymmetric) are provided by a `HybridKeyProvider`.

The primary `HybridKeyProvider` is `KeychainHybridKeyProvider`, which uses keychain queries to generate & access either RSA or ECC keys. Since `SecKey` is bound to the encryption algorithm, these keys specifically are used for ECDHC-SHA256-AES-GCM (elliptic curve to AES-GCM) and RSA-OAEP-SHA256-AES-GCM (RSA to AES-GCM). ECC keys are the suggested default since they are stored in the Secure Enclave if possible, which provides the maximum security that we can provide our users (private key data is completely inaccessible). Each provider is required a context, which is an application-specific label used to look up & store keys.

The existing `AES256CBCCipher` now implements `Cipher`, which presents a deprecation and a breaking change:
- `encrypt(data:iv:)` and `decrypt(data:)` no longer throw an `AES256CBCCipher.Error`; these have been **removed**. Instead, these functions now throw a more-generic `CryptoError` with the appropriate `CryptoError.Code`.
- `OAuth2TokenAES256CBCCipher` has been deprecated in favor of a new `OAuth2TokenCryptoCipher`, which can be provided any `Cipher` or a separate `Encryptor` and `Decryptor`.