Fix completion queue bug in ImageDownloader (#163)
### Summary
In most cases, `ImageDownloader` will be executed from the main thread. However, in scenarios where this might not be the case, we'd want to make sure the completion queue is executed in the `.current` queue (when no custom queue is specified).

### Implementation
Store `.current` queue outside of any closures to ensure the queue is the same the method was called in.