Create URLDataCache and DataDownloader (#164)
### Summary
The ImageDownloader and Cache works great but only supports images that are supported by UIImage/NSImage for the OS version of the device. If the Image doesn't support the data type, the data is lost and not cached, and we are unable to fully utilize the cache if we support other image types from our application.

`URLDataCache` and `DataDownloader` removes the Image requirement from downloading the data needed while maintaining the functionality of `ImageDownloader` and ImageCache.

### Implementation
In order to reduce duplicated code, most of the existing code from `URLImageCache.swift` and `AutoPurgingURLImageCache.swift` and has been replaced with `NSData` in place of `UIImage` and `NSImage`. This allows for more flexibility in the data types being returned from dynamic URL types that we might not always have control over.

`NSData` was chosen over `Data` in order to maintain the use of `NSCache` which requires the use of a class [NSData], instead of a struct [Data]. Since these are interoperable, I felt this wouldn't be an issue. 

### Test Plan
- Ensure Images are still downloaded, cached, and retrieved properly. This can be done using the tests or in your application. I have used Proxyman to determine that multiple requests were not being made by the application to request subsequent Images from the web. In prior versions these requests would continue to be made because once `UIImage`/`NSImage` is instantiated from `Data`, the cached data is lost.
- [New functionality] Ensure unsupported image types from the web are being stored properly. This can be done using the tests or in your application. I have used Proxyman to determine that multiple requests were not being made by the application to request subsequent Images from the web.